llm/fully-local-rag-agent-ollama-llama3.1/requirements.txt
langchain==0.2.11
tavily-python==0.3.5
langgraph==0.1.14
matplotlib==3.9.1
langchain-community==0.2.10
langchain-nomic==0.1.2
scikit-learn==1.5.1
langchainhub==0.1.20
langchain-ollama==0.1.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
rich==13.7.1